Save Expense to internal db

+++ lib/screens/add_expense.dart
@@ -17,7 +17,7 @@ class AddExpense extends ConsumerStatefulWidget {
 
 class _AddExpense extends ConsumerState<AddExpense> {
   bool _isCashIn = true;
-  var _amount = 0;
+  var _amount = 0.0;
   var _remarks = "";
   DateTime _selectedDate = DateTime.now();
   TimeOfDay _selectedTime = TimeOfDay.now();
@@ -177,7 +177,7 @@ class _AddExpense extends ConsumerState<AddExpense> {
                         return null;
                       },
                       onSaved: (value) {
-                        _amount = int.parse(value!);
+                        _amount = double.parse(value!);
                       },
                     ),
                     const SizedBox(height: 24),
